backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
python-multipart>=0.0.9,<1.0
numpy>=1.26,<3.0
requests>=2.32,<3.0
faiss-cpu>=1.8,<2.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
